// renderer/model/InfosPostMP4Series.ts
export default interface InfosPostMP4Series 
{
    tamanho: string
    tituloOriginal?: string
    tituloBrasil?: string
    criador?: string
    genero?: string
    lancamento?: string
    duracao?: string
    formato?: string
    idioma?: string
    legenda?: string
    sinopse?: string
    imgURL?: string
}